This update introduces broader unit test coverage for the collision system, including dedicated BoundingBox, BoundingCircle, and BoundingCapsule tests. It also adds fake drawable/shape test support so bounding volumes can be tested without relying on real platform rendering objects.

The collision implementation has been tightened by fixing BoundingBox issues, correcting box-to-circle moving intersection behaviour, and improving separation vector handling. Delegated collision paths now swap velocities correctly where needed, use capsule radius for capsule endpoint checks, and invert delegated separation vectors so resolution is applied in the correct direction.

Capsule support has also been improved by refactoring FakeCapsule to calculate its segment from direction-based geometry rather than rotated rectangle corners, adding a default BoundingCapsule constructor, and delegating segment-based construction through the main constructor path.

Included changes
Add ShapeMath support for rotated rectangle corner calculation.
Add fake drawable and fake shape infrastructure for bounding volume tests.
Fix FakeShape setup, constants usage, and circle radius initialisation.
Fix bugs in BoundingBox collision and separation behaviour.
Add bounding box tests.
Add bounding circle tests and tidy bounding box test names.
Fix delegated collision velocity and separation direction.
Refactor FakeCapsule geometry.
Refactor BoundingCapsule constructors.
Add bounding capsule tests.

+++ EngineInterface.NativeTests/Code/Engine/Core/GameManagerTests.cpp
@@ -154,9 +154,9 @@ namespace Engine
 
                 gm.CheckInView();
 
-                for (size_t i = 0; i < 2; i++)
+                for (int i = 0; i < 2; i++)
                 {
-                    for (size_t j = 0; j < 10; j ++)
+                    for (int j = 0; j < 10; j ++)
                     {
                         ITile* t = grid->GetTile(i, j);
                         if (t)
